Move the Method in a dedicated class
